% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{AgeDeath}
\alias{AgeDeath}
\title{Age Death}
\usage{
AgeDeath(P, population)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}
}
\description{
Males are chosen to die based on their age using a type II survival curve.
}
\keyword{death}
