% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{AssignFemale}
\alias{AssignFemale}
\title{Assign Female}
\usage{
AssignFemale(P, maleSong, femaleSong)
}
\arguments{
\item{P}{a list of parameters}

\item{maleSong}{a matrix of syllable vectors}

\item{femaleSong}{a matrix of syllable vectors}
}
\description{
Assigns females to males based on template matching.  Females are randomly chosen to pick a male from the population.  Males that better match her template are more likely to be chosen.  Assumes that there are as many females as there are males, and all birds are paired in the end.
}
\keyword{female-choice}
