% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{BasicSimulation}
\alias{BasicSimulation}
\title{Basic Simulation}
\usage{
BasicSimulation(P, freq = 1)
}
\arguments{
\item{P}{a list of parameters}

\item{freq}{how often to sample data from the simulation}
}
\description{
Runs a simulation where individual values are saved for every time step.  No parameters change during the simulation.
}
\seealso{
Other Sim Functions: \code{\link{CheckInsultPs}},
  \code{\link{CheckInvasion}}, \code{\link{GetSaveInfo}},
  \code{\link{InsultSimulation}},
  \code{\link{InvasionSimulation}},
  \code{\link{LightSimulation}},
  \code{\link{SEMSimulation}}
}
\concept{Sim Functions}
\keyword{read-write-run}
