% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{BirthDeathCycle}
\alias{BirthDeathCycle}
\title{Birth Death Cycle}
\usage{
BirthDeathCycle(P, population)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}
}
\description{
A wrapper that allows birds to die, undergo oblique learning, be born, and undergo vertical learning (in that order).
}
\keyword{birth}
\keyword{death}
