% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{CalcFractional}
\alias{CalcFractional}
\title{Calculate Consensus Fraction}
\usage{
CalcFractional(P, consensusSong)
}
\arguments{
\item{P}{a list of parameters}

\item{consensusSong}{vector of the number of tutors that sang each syllable.}
}
\description{
Calculates the probability that a bird will learn a syllable depending on how many tutors that syllable was heard from.
}
\keyword{song-learning}
