% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{CalculateAllGen}
\alias{CalculateAllGen}
\title{Calculate All Generations}
\usage{
CalculateAllGen(pa, pc, t, mAge)
}
\arguments{
\item{pa}{the proportion of adults that survive to the next age}

\item{pc}{the proportion of chicks that survive}

\item{t}{the death threshold}

\item{mAge}{the max age of the population}
}
\description{
Calculates the proportion of the population that is in each generation when the simulation starts.
}
\keyword{survival-curve}
