% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{CalculateProportion}
\alias{CalculateProportion}
\title{Calculate Adult Survival Proportion}
\usage{
CalculateProportion(n = 400, t = 1, pc = 0.3, mAge = 20)
}
\arguments{
\item{n}{the population size}

\item{t}{the death threshold}

\item{pc}{the proportion of chicks that survive}

\item{mAge}{the max age of the population}
}
\description{
Calculates the proportion of adults that survive from one timestep to the next.
}
\keyword{survival-curve}
