% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{CheckBool}
\alias{CheckBool}
\title{Check Boolean}
\usage{
CheckBool(value, valueName, NAer = FALSE)
}
\arguments{
\item{value}{user parameter to check}

\item{valueName}{name of teh checked parameter}

\item{NAer}{whther the value can be NA}
}
\description{
Checks whether a value that should be Boolean is, allowing NA if required.
}
\keyword{error-check}
