% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{CheckEncouter}
\alias{CheckEncouter}
\title{Check Encouter}
\usage{
CheckEncouter(P, learners)
}
\arguments{
\item{P}{a list of parameters}

\item{learners}{indicies of males that are alive and young enough to learn}
}
\description{
Tests whether a learner met tutors
}
\keyword{song-learning}
