% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{CheckInsultPs}
\alias{CheckInsultPs}
\title{Insult Simulation}
\usage{
CheckInsultPs(P, insultP)
}
\arguments{
\item{P}{a list of parameters}

\item{insultP}{a list of parameters to switch to at time step [when]}
}
\description{
Tests whether the initial and insult parameters are compatable.
}
\seealso{
Other Sim Functions: \code{\link{BasicSimulation}},
  \code{\link{CheckInvasion}}, \code{\link{GetSaveInfo}},
  \code{\link{InsultSimulation}},
  \code{\link{InvasionSimulation}},
  \code{\link{LightSimulation}},
  \code{\link{SEMSimulation}}
}
\concept{Sim Functions}
\keyword{error-check}
