% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{CheckInvasion}
\alias{CheckInvasion}
\title{Check Invasion}
\usage{
CheckInvasion(P, trait, stat)
}
\arguments{
\item{P}{a list of parameters}

\item{trait}{the stat to change: LrnThsh, Acc, ChanceInv, or ChanceFor}

\item{numInvader}{the number of invaders to add}

\item{invaderStat}{the stat to change: LrnThsh, Acc, ChanceInv, or ChanceFor}

\item{when}{the time step at which to introduce the insult}

\item{freq}{how often to sample data from the simulation}
}
\description{
Checks whether the noise for the [trait] was set to 0, and does so it not.  Also ensures that [stat] is within the range for the [trait]..
}
\seealso{
Other Sim Functions: \code{\link{BasicSimulation}},
  \code{\link{CheckInsultPs}}, \code{\link{GetSaveInfo}},
  \code{\link{InsultSimulation}},
  \code{\link{InvasionSimulation}},
  \code{\link{LightSimulation}},
  \code{\link{SEMSimulation}}
}
\concept{Sim Functions}
\keyword{read-write-run}
