% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{CheckMinMaxInt}
\alias{CheckMinMaxInt}
\title{Check Min/Max/Int}
\usage{
CheckMinMaxInt(value, valueName, min = 0, max = 1, maxed = FALSE,
  int = TRUE)
}
\arguments{
\item{value}{the user defined parameter}

\item{valueName}{the name of the value being checked}

\item{min}{minumum value for a feature}

\item{max}{maximum value value for a feature}

\item{maxed}{whether a value has a maximum}

\item{int}{whether a value is an integer}
}
\description{
Checks whether values that should be integers are and ensures that are within the correct range.
}
\keyword{error-check}
