% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{CheckP}
\alias{CheckP}
\title{Check Parameters}
\usage{
CheckP(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
Checks whether user defined parameters fit all requirements
}
\keyword{error-check}
