% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{CheckTrait}
\alias{CheckTrait}
\title{Check Trait}
\usage{
CheckTrait(initial, noise, min, max, name, absMax = 1)
}
\arguments{
\item{initial}{user defined initial value}

\item{noise}{user defined noise value}

\item{min}{user defined min}

\item{max}{user defined max}

\item{name}{name fo the trait in question}

\item{absMax}{absolute max possible for parameter, usually 1}
}
\description{
Checks whether user defined parameters make sense for each song-learning trait
}
\keyword{error-check}
