% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{ChooseFathers}
\alias{ChooseFathers}
\title{Choose Fathers}
\usage{
ChooseFathers(P, population, vacancy)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{vacancy}{territories that need to be filled}
}
\description{
Chooses the males who will breed.  Males must be alive and know at least one syllable.  They can be chosen locally or globally.  Males who best fit selection preferences are the most likely to father offspring.  Males can father more than one chick.
}
\keyword{birth}
