% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{ChooseTutors}
\alias{ChooseTutors}
\title{Choose Tutors}
\usage{
ChooseTutors(P, population, learners, vacancy, misc = rep(0,
  length(learners)))
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{learners}{the indicies of birds that will attempt to learn}

\item{vacancy}{the indicies of  birds that are dead}

\item{misc}{a matrix of positions of other birds that are excluded for some reason (e.g. already for consensus tutors)}
}
\description{
Randomly chooses a tutor for each learner.  Tutors must be alive, not be chicks, and must know at least one syllable.  Tutors can be chosen locally or globally.
}
\keyword{song-learning}
