% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{ClusterCalc}
\alias{ClusterCalc}
\title{Cluster Calculation}
\usage{
ClusterCalc(P, matrix)
}
\arguments{
\item{P}{a list of parameters}

\item{matrix}{a saved trait from the Basic sims (requires individual data)}
}
\description{
Calculates the cluster score of a matrix.
}
\seealso{
Other Cluster Plots: \code{\link{ClusterPlot}},
  \code{\link{GetMaxMat}}, \code{\link{QuickClusterPlot}}
}
\concept{Cluster Plots}
\keyword{stats-plotting}
