% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{Cluster Plot}
\usage{
ClusterPlot(P, trait)
}
\arguments{
\item{P}{a list of parameters}

\item{trait}{a saved trait from the Basic sims (requires individual data)}
}
\description{
Cluster Plots are normalized such that the minimal score (a smooth gradient) is zero.  The black line shows the maximal score (a checkerboard pattern) while the grey line shows the average score (no pattern).  Green line plots the score of the real data over time.  The function also prints the mean probability of getting the real values given the Min, Max, and Mean values for the matrix at each timestep.
}
\seealso{
Other Cluster Plots: \code{\link{ClusterCalc}},
  \code{\link{GetMaxMat}}, \code{\link{QuickClusterPlot}}
}
\concept{Cluster Plots}
\keyword{stats-plotting}
