% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{CreateFemaleSongs}
\alias{CreateFemaleSongs}
\title{Create Female Songs}
\usage{
CreateFemaleSongs(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
A wrapper that creates identical (uniform) or noisy female songs, and offsets their position in the syllable vector to create dialects if necessary.
}
\keyword{song-template}
