% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{DefineParameters}
\alias{DefineParameters}
\title{Define Parameters}
\usage{
DefineParameters(Rows = 20, Cols = 20, Steps = 1,
  InitialSylRepSize = 5, PrcntSylOverhang = 0.2, MaxSylRepSize = 500,
  InitialAccuracy = 0.7, InherAccuracyNoise = 0.15,
  AccuracyLimits = c(0, 1), MaxAge = 20,
  InitialLearningThreshold = 2, InherLearningNoise = 0.25,
  LearningLimits = c(0, MaxAge), InitialChancetoInvent = 0.1,
  InherChancetoInventNoise = 0, ChancetoInventLimits = c(0, 1),
  InitialChancetoForget = 0.2, InherChancetoForgetNoise = 0,
  ChancetoForgetLimits = c(0, 1), ListeningThreshold = 7,
  FatherListeningThreshold = 0.999, MinLearnedSyls = 7,
  EncounterSuccess = 0.95, LearningPenalty = 0.75, AgeDeath = TRUE,
  PrcntRandomDeath = 0.1, DeathThreshold = 1, ChickSurvival = 0.3,
  LocalBreed = FALSE, LocalTutor = FALSE, LearnerStrategy = "Add",
  ConsensusNoTut = 8, ConsensusStrategy = "Conform",
  OverLearn = FALSE, OverLearnNoTut = 3, VerticalLearnCutOff = 0.25,
  ObliqueLearning = TRUE, VerticalLearning = TRUE, RepSizePrefer = 1,
  LogScale = TRUE, MatchPrefer = 0, UniformMatch = TRUE,
  MatchScale = 1, Dialects = 1, MaleDialects = "None",
  FemaleEvolve = FALSE, ChooseMate = FALSE, SaveMatch = NA,
  SaveAccuracy = NA, SaveLearningThreshold = NA,
  SaveChancetoInvent = NA, SaveChancetoForget = NA,
  SaveNames = FALSE, SaveAge = FALSE, SaveMaleSong = FALSE,
  SaveFemaleSong = FALSE, numSim = 1000, Seed = NA)
}
\arguments{
\item{Rows}{the number of rows in the bird matrix}

\item{Cols}{the number of columns in the bird matrix}

\item{Steps}{the number of spaces away from a focal territory "local" is considered to be}

\item{InitialSylRepSize}{the number of syllables birds have a 90\% chance to know when the bird matrix is intialized}

\item{PrcntSylOverhang}{the fraction of InitialSylRepsize that birds have a 10\% and 1\% chance to know when the bird matrix is intialized}

\item{MaxSylRepSize}{the length of the syllable vector}

\item{InitialAccuracy}{the mode value for accuracy when the bird matrix is intialized}

\item{InherAccuracyNoise}{the area around the mode that can be sampled from for accuracy inheritance and establishing the initial distribution}

\item{AccuracyLimits}{the absolute min and max values that accuracy can be}

\item{MaxAge}{the maximum age of birds in the population (only in use with the type II survival curve)}

\item{InitialLearningThreshold}{the mode value for the learning threshold when the bird matrix is intialized}

\item{InherLearningNoise}{the area around the mode that can be sampled from for learning threshold inheritance and establishing the initial distribution}

\item{LearningLimits}{the absolute min and max values that the learning threshold can be}

\item{InitialChancetoInvent}{the mode value for the chance to invent when the bird matrix is intialized}

\item{InherChancetoInventNoise}{the area around the mode that can be sampled from for chance to invent inheritance and establishing the initial distribution}

\item{ChancetoInventLimits}{the absolute min and max values that the chance to invent can be}

\item{InitialChancetoForget}{the mode value for the chance to forget when the bird matrix is intialized}

\item{InherChancetoForgetNoise}{the area around the mode that can be sampled from for chance to forget inheritance and establishing the initial distribution}

\item{ChancetoForgetLimits}{the absolute min and max values that the chance to forget can be}

\item{ListeningThreshold}{the max absolute number or fraction of syllables a bird hears from one oblique tutor}

\item{FatherListeningThreshold}{the max absolute number or fraction of syllables a bird hears from his father tutor}

\item{MinLearnedSyls}{when either listenign threshodl is less than 0, this is the number of syllables a bird hears from his tutor before the fraction is applied}

\item{EncounterSuccess}{the chance that a male finds suitable tutors}

\item{LearningPenalty}{an artitrary scale for how severly longer learning is punished}

\item{AgeDeath}{whether to model death on a type II survival curve (TRUE) or random death (FALSE)}

\item{PrcntRandomDeath}{the percentage of birds that die each time step when death is random}

\item{DeathThreshold}{the numbers of birds at which a group is considered to be extinquished (you probably should not change this)}

\item{ChickSurvival}{the proportion of chicks that survive to age 1}

\item{LocalBreed}{whether empty territory are filled by chicks from local males (TRUE) or any male (FALSE)}

\item{LocalTutor}{whether obliqgue learners pick tutors from from local males (TRUE) or any male (FALSE)}

\item{LearnerStrategy}{the mode by which birds learn; can be "Add", "Forget", "AddForget", or "Consensus"}

\item{ConsensusNoTut}{the number of tutors sampled in the consensus strategy}

\item{ConsensusStrategy}{the method by which consensus decisions are made; can be "Conform" (chance = based on conformity bias), "AllNone" (all tutors must sing the syllable), "Percentage" (chance = percent of tutor that sang a syllable)}

\item{OverLearn}{whether males overlearn from many tutors as chicks}

\item{OverLearnNoTut}{the number of tutors sampled in the overlearning strategy}

\item{VerticalLearnCutOff}{this minimum value the learning window can be while still allowing males to learn vertically.}

\item{ObliqueLearning}{whether the population undergoes oblique learning (TRUE) or not (FALSE)}

\item{VerticalLearning}{whether the population undergoes vertical learning (TRUE) or not (FALSE)}

\item{RepSizePrefer}{the fraction of female preference dedicated to larger repertoires}

\item{LogScale}{whether females percieve repertoire size on a natural log scale (TRUE) or not (FALSE)}

\item{MatchPrefer}{the fraction of female preference dedicated to template matching}

\item{UniformMatch}{whether all females have the same song template (TRUE) or variations on a template (FALSE)}

\item{MatchScale}{an equation for how matching is perceived; not yet implemented!}

\item{Dialects}{the number of dialects; must be a factor of the matrix size}

\item{MaleDialects}{whether males start the simulation with dialects; can be "None" (all males are similar to dialect 1), "Similar" (male songs are in teh correct syllable space, but are not identical to female songs), "Same" (male song temapltes are identicle to their female's template)}

\item{FemaleEvolve}{whether the female templates can evolve (TRUE) or stay static throughout teh simmulation (FALSE)}

\item{ChooseMate}{whether females can pick their mate (TRUE) or not (FALSE)}

\item{SaveMatch}{whether to save matches; can be NA (the program decides based on other parameters) or TRUE/FALSE}

\item{SaveAccuracy}{whether to save the accuracy values; can be NA (the program decides based on other parameters) or TRUE/FALSE}

\item{SaveLearningThreshold}{whether to save the learning thresholds; can be NA (the program decides based on other parameters) or TRUE/FALSE}

\item{SaveChancetoInvent}{whether to save the chance to invent; can be NA (the program decides based on other parameters) or TRUE/FALSE}

\item{SaveChancetoForget}{whether to save the chance to forget; can be NA (the program decides based on other parameters) or TRUE/FALSE}

\item{SaveNames}{whether to save the UID and father's UID of the birds; can be TRUE or FALSE}

\item{SaveAge}{whether to save the age of the birds; can be TRUE or FALSE}

\item{SaveMaleSong}{whether to save male song templates; can be TRUE or FALSE}

\item{SaveFemaleSong}{whether to save female song templates; can be TRUE or FALSE}

\item{numSim}{the number of sim steps to complete}

\item{Seed}{seed to run simulation on for reproducibility}
}
\description{
Creates a parameter list and error checks chosen parameters.  This is the only place where extensive error checking is done on parameters!
}
\keyword{initialize}
