% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{EstablishDialects}
\alias{EstablishDialects}
\title{Establish Dialects}
\usage{
EstablishDialects(P, fSongs)
}
\arguments{
\item{P}{a list of parameters}

\item{fSongs}{a matrix of syllable vectors}
}
\description{
Modifies a matrix of syllable vectors to create dialects (regions of syllables that are separated from one another in the syllable space).  Regions are defined so that each dialect space is as square as possible.
}
\keyword{song-template}
