% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{FamilyTreePlot}
\alias{FamilyTreePlot}
\title{Family Tree Plot}
\usage{
FamilyTreePlot(path, byGens = TRUE)
}
\arguments{
\item{path}{path to a fodl with 2 matricies of bird UIDs.}

\item{byGens}{whether to plots the y axis by generation (TRUE) or timestep (FALSE)}
}
\description{
Experimental plot that shows which birds sired which offspring over time.  It starts from the tips, so all but one of the original lineage will be lost after enough time steps have passed.
}
\keyword{stats-plotting}
