% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{FemaleEvolve}
\alias{FemaleEvolve}
\title{Female Evolve}
\usage{
FemaleEvolve(P, population, vacancy, fatherInd)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{vacancy}{indicies of territories where male chicks were born}

\item{fatherInd}{the indicies of the male that fathered the resident male chicks}
}
\description{
Replaces females that lived on the same territory as a dead male.  New female song templates are created based on fathers that are different from the father that sired the male on her territory.  Fathers must be alive and know at least one syllable.  One created, the match between these new females and their males are recalculated.
}
\keyword{song-learning}
