% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{FinalDirections}
\alias{FinalDirections}
\title{Final Directions}
\usage{
FinalDirections(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
A wrapper that calls StepOne() and EachStep() to create lists of locality data.
}
\keyword{locality}
