% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenerateBirds.R
\name{GenerateAdultBirds}
\alias{GenerateAdultBirds}
\title{Generate Adult Birds}
\usage{
GenerateAdultBirds(P, songs)
}
\arguments{
\item{P}{a list of parameters}

\item{songs}{a matrix of syllable vectors}
}
\description{
Generates the features for each bird in the population at timestep 0.
}
\keyword{initialize}
