% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenerateBirds.R
\name{GenerateChicks}
\alias{GenerateChicks}
\title{Generate Chicks}
\usage{
GenerateChicks(P, fatherInd, territorialMales, vacancy)
}
\arguments{
\item{P}{a list of parameters}

\item{fatherInd}{the index of the fathers}

\item{territorialMales}{the population}

\item{vacancy}{the index of future chicks (aligned with fathers)}
}
\description{
Creates chicks that are similar to their fathers.
}
\keyword{birth}
