% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenerateBirds.R
\name{GenerateFounderMales}
\alias{GenerateFounderMales}
\title{Generate Founder Males}
\usage{
GenerateFounderMales(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
Creates the population at time step 0.  Generates 1) a matrix of male syllable vectors, 2) an optional matrix of female syllable vectors, 3) a dataframe of bird features, 4) an optional locality list, and 5) structures for keeping track of bird survival if a type II survival curve is implemented.
}
\keyword{initialize}
