% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{GenerateNovelSong}
\alias{GenerateNovelSong}
\title{Generate Novel Song}
\usage{
GenerateNovelSong(P, numTemplates)
}
\arguments{
\item{P}{a list of parameters}

\item{numTemplates}{the number of song templates to create}
}
\description{
Generates one or more song templates based on parameters in P; tests of inheriting [RSize0] syllables (90\%), [RSize0]*[PerROh] syllables (10\%), and [RSize0]*[PerROh] syllables (1\%), by random sampling.  Creates a numeric vector where learned syllables are 1, and unlearned syllables are 0.  It then appends 0s to the end, so the vector is of length [MaxRSize].
}
\keyword{song-template}
