% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{GetAgeGroup}
\alias{GetAgeGroup}
\title{Get Age Group}
\usage{
GetAgeGroup(P, ageRates)
}
\arguments{
\item{P}{a list of parameters}

\item{ageRates}{the output from GetAgeRates()}
}
\description{
Given the number of birds that should be in each generation, creates a vector of ages and scrambles them for random assignment.
}
\keyword{survival-curve}
