% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{GetAgeRates}
\alias{GetAgeRates}
\title{Get Age Rates}
\usage{
GetAgeRates(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
Calculates the number of birds that should be in each generation.
}
\keyword{survival-curve}
