% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{GetLearners}
\alias{GetLearners}
\title{Get Learners}
\usage{
GetLearners(P, population, vacancy)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{vacancy}{the indicies of dead birds}
}
\description{
Returns the indices of males that are alive, young enough to learn, and met tutor males.
}
\keyword{song-learning}
