% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{GetMaxMat}
\alias{GetMaxMat}
\title{Get Max Matrix}
\usage{
GetMaxMat(trait, R, C)
}
\arguments{
\item{trait}{a saved trait from the Basic sims (requires individual data)}

\item{R}{the rows in the bird matrix}

\item{C}{the columns in the bird matrix}
}
\description{
Creates a matrix that has the maximum score given the trait data.
}
\seealso{
Other Cluster Plots: \code{\link{ClusterCalc}},
  \code{\link{ClusterPlot}}, \code{\link{QuickClusterPlot}}
}
\concept{Cluster Plots}
\keyword{stats-plotting}
