% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{GetProbability}
\alias{GetProbability}
\title{Get Probability of Reproducing}
\usage{
GetProbability(P, population, usableInd)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{usableInd}{males that are alive and know at least one syllable}
}
\description{
Calculates how well a male matches female preferences for repertoires and/or matching (and/or noise, which is added uniformly to all males) to determine their probability of fathering offspring.
}
\keyword{birth}
\keyword{female-choice}
