% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{GetSaveInfo}
\alias{GetSaveInfo}
\title{Get Save Info}
\usage{
GetSaveInfo(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
Generates the information to form appropriate datastructures
}
\seealso{
Other Sim Functions: \code{\link{BasicSimulation}},
  \code{\link{CheckInsultPs}}, \code{\link{CheckInvasion}},
  \code{\link{InsultSimulation}},
  \code{\link{InvasionSimulation}},
  \code{\link{LightSimulation}},
  \code{\link{SEMSimulation}}
}
\concept{Sim Functions}
\keyword{read-write-run}
