% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{InitAgeDistribution}
\alias{InitAgeDistribution}
\title{Initial Age Distribution}
\usage{
InitAgeDistribution(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
Creates the age distribution of the population.  Either follows a type II survival curve, or uniformly samples from 1 to the max age.
}
\keyword{survival-curve}
