% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{InvasionSimulation}
\alias{InvasionSimulation}
\title{Invasion Simulation}
\usage{
InvasionSimulation(P, numInvader, trait, stat, when)
}
\arguments{
\item{P}{a list of parameters}

\item{numInvader}{the number of invaders to add}

\item{trait}{the stat to change: LrnThsh, Acc, ChanceInv, or ChanceFor}

\item{when}{the time step at which to introduce the invaders}

\item{invaderStat}{the stat to change: LrnThsh, Acc, ChanceInv, or ChanceFor}
}
\description{
Runs an invasion simulation where the time to conversion and final average value of the trait is returned.  [numInvader] invaders have their [trait] changed to [invaderStat] and their age reset to 1 time step [when].  Simulation ends when invaders are expelled, take over, or P$nSim time steps have passed.
}
\seealso{
Other Sim Functions: \code{\link{BasicSimulation}},
  \code{\link{CheckInsultPs}}, \code{\link{CheckInvasion}},
  \code{\link{GetSaveInfo}},
  \code{\link{InsultSimulation}},
  \code{\link{LightSimulation}},
  \code{\link{SEMSimulation}}
}
\concept{Sim Functions}
\keyword{read-write-run}
