% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{LearningProcess}
\alias{LearningProcess}
\title{Core Learning Process}
\usage{
LearningProcess(P, newSongs, tutorSyllables, accuracy, chanceInv)
}
\arguments{
\item{P}{a list of parameters}

\item{newSongs}{the learner's plastic song}

\item{tutorSyllables}{the syllables the learner wants to learn from the tutor}

\item{accuracy}{the learner's accuracy}

\item{chanceInv}{the learners' chance to invent}
}
\description{
Tests whether learners successfully acquire new syllables from their tutor(s) and modifies their song if this occurs.
}
\keyword{song-learning}
