% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{LearningThrshPenalty}
\alias{LearningThrshPenalty}
\title{Learning Threshold Fitness Penalty}
\usage{
LearningThrshPenalty(P, lrnThsh)
}
\arguments{
\item{P}{a list of parameters}

\item{lrnThsh}{a vector of learning thresholds in the population}
}
\description{
Calculates the fitness penalty for longer learning, which is used as the probability that a male will be chosen to die in that timestep.
}
\keyword{death}
