% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{ListeningTest}
\alias{ListeningTest}
\title{Listening Test}
\usage{
ListeningTest(P, songs, LisThrsh)
}
\arguments{
\item{P}{a list of parameters}

\item{songs}{a matrix of tutor syllable vectors}
}
\description{
Tests which syllables a learner heard from his tutor.
}
\keyword{song-learning}
