% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{LocalSearch}
\alias{LocalSearch}
\title{Local Search}
\usage{
LocalSearch(P, population, targetMale, notAvailable)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population}

\item{targetMale}{index of the territory around which local birds should be found}

\item{notAvailable}{vectors of males that cannot be chosen}
}
\description{
Given a target territory, find which local males are alive.  If none are alive, extend the search by one step.
}
\keyword{locality}
