% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{NextStepDirections}
\alias{NextStepDirections}
\title{Next Step Directions}
\usage{
NextStepDirections(currentStep, firstStep)
}
\arguments{
\item{currentStep}{the current list of location data}

\item{firstStep}{the output form OneStepDirections()}
}
\description{
Extends the locality data by one step.
}
\keyword{locality}
