% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{ObliqueLearning}
\alias{ObliqueLearning}
\title{Oblique Learning}
\usage{
ObliqueLearning(P, population, vacancy)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{vacancy}{the indicies of dead birds}
}
\description{
A wrapper that checks which birds learn, allows them to do so, then updates syllable repertoire size and match (if needed).
}
\keyword{song-learning}
