% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{OneStepDirections}
\alias{OneStepDirections}
\title{One Step Directions}
\usage{
OneStepDirections(R, C)
}
\arguments{
\item{R}{rows}

\item{C}{columns}
}
\description{
Creates the location data for what is one "step" away form each territory.
}
\keyword{locality}
