% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{OneTutorLearning}
\alias{OneTutorLearning}
\title{One Tutor Learning}
\usage{
OneTutorLearning(P, population, tutors, learners)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{tutors}{the indicies of tutor paired with each learner}

\item{learners}{the indicies of birds that will attempt to learn}
}
\description{
Allows for birds to learn from one tutor (Add, Add/Forget, or Forget strategies).  It is also called multiple times in the OverLearn strategy, where chicks add syllables from oblique tutors.
}
\keyword{song-learning}
