% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{OverLearn}
\alias{OverLearn}
\title{Over-Learn}
\usage{
OverLearn(P, population, learners)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{learners}{the indicies of birds that will attempt to learn}
}
\description{
Allows chicks to sample from tutors other than the father to add new syllables to their repertoire.
}
\keyword{song-learning}
