% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{QuickClusterPlot}
\alias{QuickClusterPlot}
\title{Quick Cluster Plots}
\usage{
QuickClusterPlot(P, path, rep = TRUE, acc = P$SAcc, lrnThsh = P$SLrn,
  match = P$SMat, chanceInv = P$SCtI, chanceFor = P$SCtF,
  age = P$SAge, AutoLayout = TRUE)
}
\arguments{
\item{P}{a list of parameters}

\item{path}{location of a folder with simulation data}

\item{rep}{whether to plot repertoire size data}

\item{acc}{whether to plot accuracy data}

\item{lrnThsh}{whether to plot learning threshold data}

\item{match}{whether to plot matching data}

\item{chanceInv}{whether to plot chance to invent data}

\item{chanceFor}{whether to plot chance to forget data}

\item{age}{whether to plot age data}

\item{autoLayout}{whether to allow the function to figure out the layout (TRUE) or not (FALSE)}
}
\description{
Creates Cluster Plotss for all saved data except MaleSongs and FemaleSongs.  See ClusterPlot for info on plot interpretation.
}
\seealso{
Other Cluster Plots: \code{\link{ClusterCalc}},
  \code{\link{ClusterPlot}}, \code{\link{GetMaxMat}}
}
\concept{Cluster Plots}
\keyword{stats-plotting}
