% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{QuickSEMPlot}
\alias{QuickSEMPlot}
\title{Quick SEM Plot}
\usage{
QuickSEMPlot(P, path, rep = TRUE, acc = P$SAcc, lrnThsh = P$SLrn,
  match = P$SMat, chanceInv = P$SCtI, chanceFor = P$SCtF,
  age = P$SAge, mSong = P$SMSng, fSong = P$SFSng,
  autoLayout = TRUE, xlab = "Time Steps", thin = 10)
}
\arguments{
\item{P}{a list of parameters}

\item{path}{location of a folder with simulation data}

\item{rep}{whether to plot repertoire size data}

\item{acc}{whether to plot accuracy data}

\item{lrnThsh}{whether to plot learning threshold data}

\item{match}{whether to plot matching data}

\item{chanceInv}{whether to plot chance to invent data}

\item{chanceFor}{whether to plot chance to forget data}

\item{age}{whether to plot age data}

\item{mSong}{whether to plot male song data}

\item{fSong}{whether to plot  female song data}

\item{autoLayout}{whether to allow the function to figure out the layout (TRUE) or not (FALSE)}

\item{xlab}{x-axis label for plot()}

\item{thin}{how often to sample a step of song data for the SongEvolve() plots; This is graphically intensive when there are a lot of syllables (default is 500), so ideally do not plot more than 100-200 time steps.}
}
\description{
A method that plots whatever data was saved in the path location.  It takes the column averages, so it works for Basic, Light, and Insult Sims, but not for Invasion Sims.  For trait plots, black lines are the average, dark grey is the inner 50% and light grey is 100% of the values.  For syllable plots, darker means that a syllable is more common.
}
\keyword{stats-plotting}
