% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{RandomDeath}
\alias{RandomDeath}
\title{Random Death}
\usage{
RandomDeath(P, population)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}
}
\description{
Randomly picks a percentage of males in the population to die.  Current age is not a relevant factor in being chosen.  Chicks are as likely to be chosen as adults.
}
\keyword{death}
