% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{ReloadParam}
\alias{ReloadParam}
\title{Reload Parameters}
\usage{
ReloadParam(filePath)
}
\arguments{
\item{filePath}{the pather where a .SEMP is located}
}
\description{
Loads a .SEMP file and converts it into a list of parameters.
}
\keyword{read-write-run}
