% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{SaveParam}
\alias{SaveParam}
\title{Save Parameterss}
\usage{
SaveParam(P, folderName, fileName = "Parameters", type = "Basic")
}
\arguments{
\item{P}{a list of parameters}

\item{folderName}{where to save the .SEMP}

\item{type}{the simulation type run (accepts any string)}

\item{fName}{file name for the .SEMP}
}
\description{
Saves a list of parameters as a .SEMP file.
}
\keyword{read-write-run}
