% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{SEMSimulation}
\alias{SEMSimulation}
\title{SEM Simulation}
\usage{
SEMSimulation(P, type = "Basic", folderName = NA, save = TRUE,
  return = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{P}{a list of parameters}

\item{type}{what type of simulation to run ('Basic', 'Light', 'Insult)}

\item{folderName}{where to save the simualtion data, defaults to a timestemp in the current directory}

\item{save}{whether to write the data to .csvs}

\item{return}{whether to return the data in R}

\item{verbose}{whether to print the running time and folder name}

\item{...}{arguments to the simulation types.  See documentation for individual sim type arguments.}
}
\description{
A wrapper that conveniently handles data saving and times the simulation
}
\examples{
P <- DefineParameters(RepSizePrefer = 0, MatchPrefer = 1, numSim=100)
SEMSimulation(P, 'Basic', 'Example', return=TRUE)
SEMSimulation(P, 'Interval', 'Example', return=TRUE, freq=2)

P2 <- DefineParameters(numSim=600, MatchPrefer = 1, RepSizePrefer = 0)
P3 <- DefineParameters(numSim=600, SaveMatch = TRUE)
SEMSimulation(P2, insultP=P3, 'Insult', when=100, freq=2, save=FALSE, return = TRUE)
}
\seealso{
Other Sim Functions: \code{\link{BasicSimulation}},
  \code{\link{CheckInsultPs}}, \code{\link{CheckInvasion}},
  \code{\link{GetSaveInfo}},
  \code{\link{InsultSimulation}},
  \code{\link{InvasionSimulation}},
  \code{\link{LightSimulation}}
}
\concept{Sim Functions}
\keyword{read-write-run}
