% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{SongPlot}
\alias{SongPlot}
\title{Song Plot}
\usage{
SongPlot(P, songs, thin = 10, male = TRUE)
}
\arguments{
\item{P}{a list of parameters}

\item{songs}{male or female song data from  simulation}

\item{thin}{how often to sample a step of song data for the SongEvolve() plots; This is graphically intensive when there are a lot of syllables (default is 500), so ideally do not plot more than 100-200 time steps.}

\item{male}{whether male songs are being plotted; affects the y-axis label}
}
\description{
Shows the prevalence of each syllable across time.  Darker color means that a syllable is more common.
}
\keyword{stats-plotting}
