% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{TestLearningThreshold}
\alias{TestLearningThreshold}
\title{Test Learning Threshold}
\usage{
TestLearningThreshold(P, males)
}
\arguments{
\item{P}{a list of parameters}

\item{males}{the bird trait data.frame from the population of birds ($Males)}
}
\description{
Tests whether males are young enough to learn.
}
\keyword{song-learning}
