% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_StatsPlotsandTesting.R
\name{TraitPlot}
\alias{TraitPlot}
\title{Trait Plot}
\usage{
TraitPlot(trait, xlab = "Time Steps", ylab)
}
\arguments{
\item{trait}{a trait matrix to plot}

\item{xlab}{x-axis label for plot()}

\item{ylab}{y-axis label for plot()}
}
\description{
Plots averages for a bird trait.  Black lines are the average, dark grey is the inner 50% and light grey is 100% of the values.  For syllable plots, darker red means that a syllable is more common.
}
\keyword{stats-plotting}
