% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_LifeDeathFunctions.R
\name{UpdateProbabilities}
\alias{UpdateProbabilities}
\title{Update Survival Proportions}
\usage{
UpdateProbabilities(P, chicks, prob)
}
\arguments{
\item{P}{a list of parameters}

\item{chicks}{vector of chick indicies}

\item{prob}{the data structure from the population that keeps track of survival probabilities}
}
\description{
At the end of a timestep, it removes the adult survival proportion from the generation that has just been extinguished, and calculates the  adult survival proportion for the chicks that have just been generated.
}
\keyword{death}
\keyword{survival-curve}
