% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{UpdateSongTraits}
\alias{UpdateSongTraits}
\title{Update Song Traits}
\usage{
UpdateSongTraits(P, population, learners)
}
\arguments{
\item{P}{a list of parameters}

\item{population}{the population of birds}

\item{learners}{the indicies of birds that will attempt to learn}
}
\description{
Updates the SylRep and Match traits for learners post learning.
}
\keyword{song-leanring}
