% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Learning.R
\name{VerticalSongLearning}
\alias{VerticalSongLearning}
\title{Vertical Song Learning}
\usage{
VerticalSongLearning(P, templates, chicks)
}
\arguments{
\item{P}{a list of parameters}

\item{templates}{matrix of the fathers' syllable vectors}

\item{chicks}{song-learning traits of chicks form the population ($Males)}
}
\description{
A wrapper that prepares chick and tutor template data during vertical learning.
}
\keyword{song-learning}
