% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Parameters_and_Simulation.R
\name{LightSimulation}
\alias{LightSimulation}
\title{Light Simulation}
\usage{
LightSimulation(P, freq = 200)
}
\arguments{
\item{P}{a list of parameters}

\item{freq}{how often to sample data from the simulation}
}
\description{
Runs a simulation where the only average values are saved every [freq] time step.  No parameters change during the simulation.
}
\seealso{
Other Sim Functions: \code{\link{BasicSimulation}},
  \code{\link{CheckInsultPs}}, \code{\link{CheckInvasion}},
  \code{\link{GetSaveInfo}},
  \code{\link{InsultSimulation}},
  \code{\link{InvasionSimulation}},
  \code{\link{SEMSimulation}}
}
\concept{Sim Functions}
\keyword{read-write-run}
